kafka-app/requirements.txt
requests
psycopg2-binary
kafka-python
python-dotenv
